==> 10 Selenium, Projekte/projektbsp/03 Analysis CSV.ipynb <==
import pandas as pd
from ast import literal_eval

import matplotlib.pyplot as plt
import matplotlib
plt.style.use('fivethirtyeight')
%matplotlib inline

!ls

#Importing the data
df = pd.read_csv('readable_cleaned.csv')
del df['Date.1']
df.index = pd.to_datetime(df['Date'], format='%Y-%m-%d %H:%M:%S')

df.info()

# Distribution over time

## Weekly

# Distribution
df[df['Date']!= '2001-01-01 01:01:01'].resample('W')['Number'].count().plot()

df[df['Date']!= '2001-01-01 01:01:01'].resample('W')['Number'].count().sort_values(ascending=False).head()

## Daily

df[df['Date']!= '2001-01-01 01:01:01'].resample('D')['Number'].count().plot()

df[df['Date']!= '2001-01-01 01:01:01'].resample('D')['Number'].count().sort_values(ascending=False).head()

## Most active Day

df['2016-05-12'].sort_values(by='Date')

df['Number'].groupby(df['Number'].index.weekday).count()

## Most active time of day

df['Number'].groupby(df['Number'].index.hour).count()

# Ad Click Count

df['Ad Clicks'].sum()

# Ad click Distribution

df['Ad Clicks'].describe()

# Ad Click top 5

df['Ad Clicks'].sort_values(ascending=False).head()

pd.set_option('max_colwidth',400)

df[df['Ad Clicks']>40200]

# Ad Impressions

df['Ad Impressions'].sum()

df['Ad Impressions'].describe()

df['Ad Impressions'].sort_values(ascending=False).head()

df[df['Ad Impressions']>5112241

## Ad Spend

df['Ad Spend'].sum()

df['Ad Spend'].sort_values(ascending=False).head()

# Age
## Percentage were targetted at under age 

df[df['AgeMin']< 18]['Ad Impressions'].sum()

df[df['AgeMin']< 18]['Ad Impressions'].sum() / df['Ad Impressions'].sum() * 100

## Interests

lst = list(df['Interests'])
lst[0]

new_lst = []
for elem in lst:
    new_lst.append(literal_eval(elem))
flat_list = [item for sublist in new_lst for item in sublist]
fl = []
for elem in flat_list:
    elem = elem.upper().replace(' OR', '').strip()
    fl.append(elem)
df_tags = pd.DataFrame(fl)
df_tags.columns = ['Tags']
df_tags['Tags'].count()

df_tags['Tags'].value_counts().head(40)
df_tags = df_tags[df_tags['Tags']!='N/A']
df_tags = df_tags[df_tags['Tags']!='JR.']
df_tags = df_tags[df_tags['Tags']!='']
df_tags['Tags'].value_counts().head(15).sort_values(ascending=True).plot(kind='barh')

#After election
lst = list(df['2017-05-01':]['Interests'])
new_lst = []
for elem in lst:
    new_lst.append(literal_eval(elem))
flat_list = [item for sublist in new_lst for item in sublist]
fl = []
for elem in flat_list:
    elem = elem.upper().replace(' OR', '').strip()
    fl.append(elem)
df_tags = pd.DataFrame(fl)
df_tags.columns = ['Tags']
df_tags['Tags'].value_counts().count()
df_tags['Tags'].value_counts().head(40)

# Location

!ls 

df_usstates = pd.read_csv('US.csv')
states = list(df_usstates['name'])

statecounts = []

for state in states:
    new = pd.DataFrame(df['Location'].str.contains(state))
    count = new[new['Location']==True].count()[0]
    mini_dict = {'State':state,
                 'Count':count}
    statecounts.append(mini_dict)

pd.DataFrame(statecounts).to_csv('state_count.csv')

df_usstates = pd.DataFrame(statecounts)

df_usstates.sort_values(by='Count', ascending=False).head(10)

poly = pd.read_csv('cartodb_query.csv')

merged = df_usstates.merge(poly, left_on='State', right_on='name')

merged.to_csv('state_count_with_polygon.csv')

## Finding white nationalism

def unlist(elem):
    elem = str(elem)
    return elem
df['Interests'] = df['Interests'].apply(unlist)    

df['White'] = df['Interests'].str.contains('illegal')

df[df['Interests']=="['Confederate Flag', 'American Civil War reenactment']"]

df['Interests']
